Fixes #13811 -  Refactor RunHostJob and RunProxyCommand to support more providers
The current code for RunHostJob and RunProxyCommand is effectively just
the code to run SSH. Most of the code can be reused in other providers,
by creating subclasses RunSSHJob and RunProxySSHCommand which only
contain the specific bits for the SSH provider.
