+++ app/lib/actions/remote_execution/run_host_job.rb
@@ -1,7 +1,6 @@
 module Actions
   module RemoteExecution
     class RunHostJob < Actions::EntryAction
-
       middleware.do_not_use Dynflow::Middleware::Common::Transaction
       include Actions::RemoteExecution::Helpers::LiveOutput
 
@@ -10,32 +9,16 @@ module Actions
       end
 
       def plan(job_invocation, host, template_invocation, proxy)
-        action_subject(host, :job_category => job_invocation.job_category, :description => job_invocation.description)
-
-        template_invocation.host_id = host.id
-        template_invocation.run_host_job_task_id = task.id
-        template_invocation.save!
-
+        save_template_invocation(host, template_invocation)
+        validate_template_invocation!(template_invocation)
+        validate_proxy!(proxy)
         link!(job_invocation)
         link!(template_invocation)
+        action_subject(host,
+                       :job_category => job_invocation.job_category,
+                       :description  => job_invocation.description)
 
         verify_permissions(host, template_invocation)
-        hostname = find_ip_or_hostname(host)
-
-        raise _('Could not use any template used in the job invocation') if template_invocation.blank?
-
-        settings = { :global_proxy   => 'remote_execution_global_proxy',
-                     :fallback_proxy => 'remote_execution_fallback_proxy' }
-
-        raise _('Could not use any proxy. Consider configuring %{global_proxy} ' +
-                'or %{fallback_proxy} in settings') % settings if proxy.blank?
-
-        renderer = InputTemplateRenderer.new(template_invocation.template, host, template_invocation)
-        script = renderer.render
-        raise _('Failed rendering template: %s') % renderer.error_message unless script
-
-        provider = template_invocation.template.provider
-        plan_action(RunProxyCommand, proxy, hostname, script, provider.proxy_command_options(template_invocation, host))
         plan_self
       end
 
@@ -68,20 +51,6 @@ module Actions
         N_('Remote action:')
       end
 
-      def find_ip_or_hostname(host)
-        %w(execution primary provision).each do |flag|
-          if host.send("#{flag}_interface") && host.send("#{flag}_interface").ip.present?
-            return host.execution_interface.ip
-          end
-        end
-
-        host.interfaces.each do |interface|
-          return interface.ip unless interface.ip.blank?
-        end
-
-        return host.fqdn
-      end
-
       private
 
       def verify_permissions(host, template_invocation)
@@ -94,6 +63,27 @@ module Actions
         raise _('User can not execute this job template on %s') % host.name unless authorizer.can?(:execute_template_invocation, template_invocation)
 
         true
+      end
+
+      def save_template_invocation(host, template_invocation)
+        template_invocation.host_id = host.id
+        template_invocation.run_host_job_task_id = task.id
+        template_invocation.save!
+      end
+
+      def validate_template_invocation!(template_invocation)
+        raise ::Foreman::Exception.new(
+          N_('Could not use any template used in the job invocation')
+        ) if template_invocation.blank?
+      end
+
+      def validate_proxy!(proxy)
+        settings = { :global_proxy   => 'remote_execution_global_proxy',
+                     :fallback_proxy => 'remote_execution_fallback_proxy' }
+        raise ::Foreman::Exception.new(
+          N_('Could not use any proxy. Consider configuring %{global_proxy} ' +
+             'or %{fallback_proxy} in settings') % settings
+        ) if proxy.blank?
       end
     end
   end
