+++ app/lib/actions/remote_execution/run_proxy_command.rb
@@ -1,18 +1,8 @@
 module Actions
   module RemoteExecution
     class RunProxyCommand < Actions::ProxyAction
-
       include ::Dynflow::Action::Cancellable
       include Actions::RemoteExecution::Helpers::LiveOutput
-
-      def plan(proxy, hostname, script, options = {})
-        options = { :effective_user => nil }.merge(options)
-        super(proxy, options.merge(:hostname => hostname, :script => script))
-      end
-
-      def proxy_action_name
-        'Proxy::RemoteExecution::Ssh::CommandAction'
-      end
 
       def on_data(data)
         if data[:result] == 'initialization_error'
@@ -24,7 +14,6 @@ module Actions
           error! _('Script execution failed') if failed_run?
         end
       end
-
 
       def rescue_strategy
         ::Dynflow::Action::Rescue::Skip
