+++ app/lib/actions/remote_execution/ssh/run_proxy_command.rb
@@ -1,0 +1,16 @@
+module Actions
+  module RemoteExecution
+    module SSH
+      class RunProxyCommand < RunProxyCommand
+        def plan(proxy, hostname, script, options = {})
+          options = { :effective_user => nil }.merge(options)
+          super(proxy, options.merge(:hostname => hostname, :script => script))
+        end
+
+        def proxy_action_name
+          'Proxy::RemoteExecution::Ssh::CommandAction'
+        end
+      end
+    end
+  end
+end
