+++ test/unit/actions/run_hosts_job_test.rb
@@ -43,7 +43,7 @@ module ForemanRemoteExecution
     end
 
     it 'triggers the RunHostJob actions on the resolved hosts in run phase' do
-      planned.expects(:trigger).with() { |*args| args[0] == Actions::RemoteExecution::RunHostJob }
+      planned.expects(:trigger).with() { |*args| args[0] == Actions::RemoteExecution::SSH::RunJob }
       planned.create_sub_plans
     end
